server/app/requirements.txt
accelerate==0.15.0
anyio==3.6.2
black==22.10.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
clip==1.0
contourpy==1.0.6
cycler==0.11.0
diffusers==0.9.0
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.88.0
filelock==3.8.2
fonttools==4.38.0
ftfy==6.1.1
h11==0.14.0
httpcore==0.16.2
httptools==0.5.0
httpx==0.23.1
huggingface-hub==0.11.1
idna==3.4
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
mypy-extensions==0.4.3
numpy==1.23.5
opencv-python==4.6.0.66
orjson==3.8.3
packaging==22.0
pathspec==0.10.2
Pillow==9.3.0
platformdirs==2.6.0
psutil==5.9.4
pydantic==1.10.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
rfc3986==1.5.0
semantic-version==2.10.0
setuptools-rust==1.5.2
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tokenizers==0.13.2
tomli==2.0.1
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
transformers @ git+https://github.com/huggingface/transformers@9e56aff58a742b48fc8edea8d28d5b80330efbcc
clip @ git+https://github.com/openai/CLIP.git
typing_extensions==4.4.0
ujson==5.6.0
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
wcwidth==0.2.5
websockets==10.4
zipp==3.11.0
